requirements.txt
phidata
python-dotenv
fastapi
uvicorn
groq
opencv-python
pytesseract
numpy
pdf2image
pillow
pandas
